Add a Layout to a new api
Currently I'm putting the api under api/.  Aleksa had also suggested
"github.com/openSUSE/umoci/umoci", but that conflicts with the final
built binary and is in .gitignore.  I assume we'll be moving it, but
for now just stick it under github.com/openSUSE/umoci/api to get the
ball rolling.

This only introduces a Layout object.  I don't like 'OpenLayout' as
a call for opening it, but again it's just a start.

Signed-off-by: Serge Hallyn <shallyn@cisco.com>